Try move to data approach (#30)
- implement raws files (via .jsonc)
- camera zoom
- add a bunch of additional items
- improve pathfinding code

+++ game_defence/src/modules/camera/helpers.hpp
@@ -6,7 +6,7 @@
 
 namespace game2d {
 
-[[nodiscard]] glm::ivec2
+[[nodiscard]] glm::vec2
 mouse_position_in_worldspace(entt::registry& r);
 
 [[nodiscard]] glm::ivec2
